users-service/requirements.txt
Flask
psycopg2-binary
redis
prometheus_client
python-dotenv
marshmallow
Flask-Limiter
Flask-Cors